requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=1.10.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
PyYAML>=5.4.0
wandb>=0.12.0
scipy>=1.7.0
